1) Changed both IgnoreMinerals checkboxes to tristate (all/low-sec only/no ore). 2) Changed all paths to point to %appdata%\EVEIPH 3) Changed all path generations to use Path.Combine instead of plain string concats to remove separator issues. 4) Fixed negative id's in GetAssetLocationAndFlagInfo call. 5) Fixed an exception issue with periodic updating in frmIndustryJobsView.

+++ EVEAssets.vb
@@ -273,7 +273,8 @@ Public Class EVEAssets
 
                 If readerData.Read Then
                     ' Call this function again to get the location name
-                    LocationName = GetAssetLocationAndFlagInfo(readerData.GetInt32(0), FlagID, FlagText)
+                    ' veg fixed reading negative id values.
+                    LocationName = GetAssetLocationAndFlagInfo(readerData.GetInt64(0), FlagID, FlagText)
                     readerData.Close()
                 Else
                     ' See if it's a Citadel
